requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
cbor==1.0.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==3.6.0
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docx2txt==0.9
durationpy==0.9
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
FlagEmbedding==1.3.4
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
ijson==3.4.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inscriptis==2.6.0
ir_datasets==0.5.10
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.25
langchain-chroma==0.2.3
langchain-community==0.3.24
langchain-core==0.3.59
langchain-huggingface==0.2.0
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.39.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.8
onnxruntime==1.19.2
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
peft==0.15.2
pillow==9.2.0
posthog==4.0.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pypdf==5.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
trec-car-tools==2.6
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.4.0
uvicorn==0.34.2
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zlib-state==0.1.9
zstandard==0.23.0
setuptools
wheel
torch @ file:///C:/Users/I-JINGHUAN.JH.WU/Desktop/chatbot-venv/wheelhouse/torch-2.1.2+cu121-cp39-cp39-win_amd64.whl